rdkit dependency solved
