--- ersilia/setup/requirements/rdkit.py
@@ -1,15 +1,0 @@
-import sys
-import importlib
-from ...utils.terminal import run_command
-
-
-class RdkitRequirement(object):
-    def __init__(self):
-        self.name = "rdkit"
-        try:
-            importlib.import_module(self.name)
-        except:
-            self.install()
-
-    def install(self):
-        run_command("python -m pip install rdkit-pypi")
