+++ requirements.txt
@@ -7,7 +7,6 @@ joblib
 hashids
 bioservices
 biopython
-chembl_webresource_client
 markdown
 pycrypto
 python-dateutil<2.8.1,>=2.1
